Enhanced putting in height data and cleaned
made so it automatically formats the height as well as is deletable, was running into issue with loop to keep it in a format was fighting back with backspace, also got rid of unneccsarry packages folder. next is to clean the code and get ride of pixels being over flwoed.